requirements.txt
Flask
fitz
PyMuPDF
python-docx
openai